+++ tests/testthat/test-tidyeval.R
@@ -56,6 +56,21 @@ test_that("translate context functions", {
   expect_equal(capture_dot(dt, cur_group_rows()), quote(.I))
 })
 
+test_that("translates case_when()", {
+  dt <- lazy_dt(data.frame(x = 1:10, y = 1:10))
+
+  expect_equal(
+    capture_dot(dt, case_when(x1 ~ y1, x2 ~ y2)),
+    quote(fcase(x1, y1, x2, y2))
+  )
+
+  # translates recursively
+  expect_equal(
+    capture_dot(dt, case_when(x == 1 ~ n())),
+    quote(fcase(x == 1, .N))
+  )
+})
+
 test_that("can process many expressions in one go", {
   dt <- lazy_dt(data.frame(x = 1:10, y = 1:10))
   n <- 10